Déplacement de Tactic.py et Action.py dans ai
